requirements.txt
langchain_openai
streamlit
langchain
langchain_community
mysql-connector-python
python-dotenv